Set NEWSC bit when sending probe config
This bit also needs to be set so that the gateware module starts sending
the config data to the ASIC. The setting of this bit is done is done in
the citiroc_send_probes() function [citiroc.c].

Additionally, this commit also uncomments the calls to the folowing two
functions in main.c:

-- citiroc_send_config()
-- citiroc_send_probes()

While the config is received correctly at the MSS ebd, the ASIC still
does not respond to the config we're sending it. To be tested why.