Changes for Citiroc configuration
The cfg_ram now contains the slow control and the probes data at
different memory offsets. This has been reflected in the mem_mgmt files.

Also, this commit adds a new take on the writing to the cfg_ram after
the data is received from the OBC siulator.

Finally, the calls to citiroc_send_slow_control() and
citiroc_send_probes() have been uncommented, to allow the configuration
to be sent to the ASIC.

The ASIC still does not respond to the configuration it is applied
through the slow control and probes configuration data. To be examined
further why this is.