Add command for setting DAQ time and changing command names
The command names were edited to comply with the MSP naming standards
(the unofficial ones): send commands should have a "SEND" text in them.

Another fairly important change made was to clear the current setting of
ROCSR.DAQDUR before applying the new setting. Otherwise, an OR-ing of
bits could happen, yielding to an incorrect setting of DAQDUR.