Add command to set temporary HVPS voltage
The voltage is set using the HBV command of the C11204-02 HVPS module.
A new MSP command (MSP_OP_SEND_HVPS_TMP_VOLT) is implemented, that
sends two bytes of voltage setting over MSP. The two bytes are passed to
the new hvps_set_temporary_voltage() function, which accepts a uint16_t
that it converts to the ASCII value expected by the HVPS as part of the
HBV command.