Change mem_read() to actually copy bytes and return num. bytes written
It was previously not writing bytes, only copying the pointer address of
the memory location to the "data" parameter. With the modifications in
this commit, the function copies bytes to the data parameter and returns
the number of bytes written.

Changes applied in start_hvps() function (hvps_c11204.c file) - BUT not
tested fully.