Add HGS command to retrieve HVPS status
The HVPS driver now also sends the HGS command on the Timer1_IRQ. The
command obtains the status of the HVPS and stores it to the hvps_status
variable (local to the hvps_*.c file).

A function to retrieve the ON/OFF status of the HVPS [hvps_is_on()] is
also included as part of this commit.

NB: Code compiles, but is untested.