Change HVPS data retrieval policy
1. Retrieve data via commands every second; on every second, another
type of data is obtained. The cycle is the following:

a. HVPS status
b. HVPS output voltage
c. HVPS output current
d. HVPS output temperature
e. Repeat from (a)

2. Remove calls to timer start and stop functions in HVPS turn on and
off, respectively. We want data from the HVPS to be retrieved even when
it is off.

3. Remove the hvps_hk_set_timer() function. It was redundant, PLUS we're
running out of stack memory on the processor (woo-hoo!), so some
unnecessary things have to go.