Remove clearing of MSP HK buffer
With the way it was previously implemented, the clearing of the HK
buffer yielded zero readouts from time to time in the CitirocUI window -
and in any form of readout that would be used.

In the case of telemetry in particular, with the way the HVPS telemetry
is currently being read - on five-second intervals - any MSP HK readout
performed less than five seconds apart would yield all-zeroes on the
HVPS telemetry with the MSP HK buffer clearing. Confusions such as this
should be avoided.