Add reset counter functionality
Adds functions that increments a NVM counter whenever the system is
booted.

Has a read function which adds the uint32_t value to the end of
housekeeping data.

Has a reset function which sets the counter to 0 and is called by the
OBC as a syscommand with OPcode 0x55.

NB: These changes are untested.