Attemt unsuccessfully to avoid HardFault in hvps_get_* fcns
Functions like hvps_get_status(), hvps_get_voltage() result in a
HardFault handler. This commit attempts unsuccessfully to fix this
by casting some unsigned char pointers into signed char pointers.

Also, the incrementing of current_run in the UART handler has been moved
to the end of the function. This results in the status always being
obtained first, and in the case of the problem at hand, it simply
confirms that not just the hvps_get_voltage() function puts us in
HardFault handler.