Fix BX introduced by compiler and clean up HVPS code
The BX was likely introduced by the following line in the previously
implemented "getarray()" function:

	memset(cmd, '\0', 28);

In particular, the 28 part at the end, which was larger than the "cmd"
array, was messing up the memory to such degree that it probably made
the compiler think it should introduce a BX instruction to an address
that did not exist. This BX resulted in a HardFault being introduced in
the processor. A HardFault is (usually) the embedded equivalent of a
Linux seg-fault, so it all makes sense.

This commit also changes the "getarray()" function into one which is
more aptly named, and contains a single parameter. This highlights its
actual use better.

Apart from this, most uint8_t pointerss were changed to char pointers,
to avoid compiler warnings; anyway, the HVPS expects ASCII chars as
command strings, so char pointers better reflect this usage scenario
than uint8_t pointers.