Add support for DAQ temperature header
Adds functions for reading out the temperature, current and voltage
before a DAQ starts and before it ends. This is done by timer64 for
housekeeping being split up to timer1 and timer2. Timer1 is kept
for housekeeping, while timer2 is used for DAQ counting.
Citiroc functions are called when this triggers to store values in
gateware registers.