Fix potential seg-fault in new HVPS functions
This commit allows for DAQ to run properly with the HVPS TM written to
the HVPSR and subsequently (by gateware) to the histogram header.

The fix is to increase the value of the array used to store HVPS
voltage, current and temperature from four to 8. Since strtol is used on
the array in each function, I was thinking that the length of the array
is not taking into account for C string terminator (\0). However, making
the array five bytes long and initializing it to all zeros still leads
to the error.

Hence, I think the problem may be hiding somewhere else. It's very
interesting that a HardFault is generated _somewhere_ in code, which
leads to an MSP error getting triggered.

In short, pehaps it's time to look more deeply into _all_ array lengths
used and what is being written to arrays. For starters, code outside MSP
should be checked. If nothing is found there, one should start looking
in MSP, too.