Remove applying configurations from NVM
In order to avoid misconfiguring the HVPS and ASIC due to missing
configuratoin in the NVM, the code used to apply default HVPS and ASIC
config's has been removed. To be decided later on whether it will be
reintroduced.

In hvps_c11204-02.c, the code has also been simplified by removing two
functions called only once (thus not necessary).