Make HVPS readout work
With the new implementation, the UART interrupt handler on the HVPS
response was not executing. The issue was because with the new
implementation, an I2C command is used to initiate a readout of the HVPS
data. It turns out that the I2C interrupt does not finish exeuction,
seemingly due to the MSP callbacks still executing. Because of the I2C
interrupt still being active, the UART interrupt's handler cannot be
executed, since the UART interrupt has the same priority as the I2C.
Making the I2C interrupt a lower priority than the UART fixes the
problem, enabling the UART handler to pre-empt the I2C's handler.