Make uniform use of tab and space characters
In most files, tab characters have been used to align code contexts,
e.g., the contents of an `if` or a `for` statement, or code which is
part of functions. Spaces have been used to align text, for example to
align an addition statement prior to its assignemnt to a value in a
statement which goes beyond 80 columns.

In the few files (including those generated by Microsemi tools) that use
spaces instead of tabs for aligning code within a context, the spaces
have been left untouched.

Closes #40.