Add stubs for two new HVPS functions
The functions in question are `hvps_get_status()` and
`hvps_get_last_cmd_err()`. The former will retrieve the status using the
`HGS` command and the latter will return the status field of the last
`hxx` command (if any) obtained from the HVPS module.

For now, as the title of the commit message states, these functions are
stubs: they only return 0.