Remove logic to store MSP seqflags every 16th transaction
The logic doesn't really make sense. There is no particular periodicity
to MSP commands sent to CUBES, so storing the flags once every 16th
transaction doesn't necessarily mean the flags would be synchronized. In
addition, an unexpected power off could occur sometime between the
storage of MSP seq flags, which would mean the sequence flags would
anyway be de-synchronized.

Storing the flags to NVM on every transaction is unfortunately not
feasible on CUBES, since our NVM is flash with around max. 10k write
cycles. With the way CUBES is currently intended to be operated on
orbit, there are at least REQ_PAYLOAD commands being sent to CUBES once
every 10 s. Considering just these and the 10k write cycles per page
available on the flash, we would reach the max. number of write cycles
in 27.8 hours:

10,000 cycles * 10 s (cmd. periodicity) = 100,000 s to reach 10k cycles
100,000 s / 3600 s = 27.777 hrs. to reach 10 k cycles

This is only considering REQ_HK commands; obviously, there would be
more commands sent to CUBES within its (planned) 24-hour operation
period every time it is turned on (CUBES will operate multiple times
during the planned one year lifetime of MIST; every time it is operated,
it will be operated for 24 hours). The page of the flash memory
containing the sequence flags would be almost be worn out during one
operational period. This is of course unacceptable, so the decision has
been made to only store the sequence flags on POWER_OFF, as requested by
MIST/MSP.

If an unexpected reboot does occur, what would happen is either that a
command will be aborted and retried by the OBC, or worst-case duplicate
data will be retrieved from CUBES. The latter can be lived with (for
now).

Closes #50.