Change LED interface
The `led_blink` function was changed to `led_blink_for` and it now
blinks the LED only once, for a certain number of ms. This allows for
using this function in debug more easily. A new function,
`led_blink_repeat` was also added and can be used to repeatedly blink
the LED with a specified ms delay.