Remove unnecessary msp_i2c files
These files were part of the generated msp folder. As part of the
migration of the MSP experiment functions to `main.c`, the issue of
manually-written files added to those in the `msp/` folder, which
contains only generated files, was solved. The contents of these files
were moved to `main.c`.