Read HK once a second
In order for CUBES to not clock-stretch while it is reading HK
(something that can be a lengthy process due to the HVPS being
communicated to several over UART at a speed ten times lower than the
I2C), I made the decision of reading HK once a second. For this purpose,
the Timer64 was reverted to a 32-bit timer (using Timer1) and it
triggers once a second. When DAQ is enabled, a counter is used to count
down from DAQDUR for DAQDUR-1 seconds, which is used to prepare the HK
for storage to histogram headers during DAQ.