Have another take at blinking in HardFault_Handler
The blinking was observed not to be working, because the led_blink_for()
function uses the SysTick_Handler() and the SysTick interrupt has a
lower priority than HardFault, which means it cannot preempt
HardFault_Handler() until it has finished.

This commit attempts to fix this by "manually" blinking the LED, which
should hopefully work. It could not yet be tested properly, as HardFault
is not really so easy to achieve...