De-inhibit prep_payload_data() also on DAQ stop
... and adding comments to function prototypes at the top. This was
actually an example of a comment saving the life of some data files (at
least). It turned out that, following the latest changes, the data files
on "forced" DAQ stop (such as via MSP_OP_CUBES_DAQ_STOP or on power-off)
were not actually saved! This commit should address that.