Read HVPS temp. compensation correctly
There was a hidden bug related to using `strtol()` for reading out the
temperature compensation: `strtol()` reads "as much as it can" of the
input string and attempts to convert. Because the HVPS module replies to
the `HRT` command, the calls to `strtol()` from
`hvps_get_temp_comp_factor()` lead to wrong values being written to the
output temp. corr. factor `struct`.

In this commit, I make sure we read only the bytes actually needed for
each conversion to string and run `strtol()` on those bytes.

I also adjusted the data types in the `struct hvps_temp_corr_factor`, as
`dTp1` and `dTp2` are actually signed integers. `strtol()` is then used
for only these two parameters and `strtoul()` is used for the `uint8_t`.