Address warning on call to mem_nvm_write()
There was a warning (that was at times missed by the compiler) on the
call to mem_nvm_write() from the code executing upon the reception of a
MSP_OP_SEND_NVM_CITI_CONF. The parameter used was the same nvm_conf_addr
used for checking whether a valid configuration exists. The problem was
that `nvm_conf_addr` is a `uint8_t *`, while mem_nvm_write() expects a
`uint32_t` as the first parameter. Hence, the warning -- and the
potential for disastrous error.

Also done in this commit: added some TODOs in case anyone in the future
wants to check that all bytes in the MSP commands for sending Citiroc
conf's to CUBES.