Clean up for Release to NVM build
The Release build will actually be a Debug build with the linker script
changed to "debug in ENVM". That is the only way that was observed to be
working, it is not a problem to use it, so it will be used.

The Release build has been removed from the Eclipse .cproject file, and
the Release launch configuration and other unnecessary linker scripts
imported from Libero have been removed.